// packages/graphql/src/__generated__/graphql.ts
export type GetProductDetailForProductPageQuery = {
    products: {
        items: Array<
            | ({ id: number | null; uid: string } & {
                  " $fragmentRefs"?: {
                      ProductDetailsFragment_BundleProduct_Fragment: ProductDetailsFragment_BundleProduct_Fragment;
                  };
              })
            | ({ id: number | null; uid: string } & {
                  " $fragmentRefs"?: {
                      ProductDetailsFragment_ConfigurableProduct_Fragment: ProductDetailsFragment_ConfigurableProduct_Fragment;
                  };
              })
            | ({ id: number | null; uid: string } & {
                  " $fragmentRefs"?: {
                      ProductDetailsFragment_DownloadableProduct_Fragment: ProductDetailsFragment_DownloadableProduct_Fragment;
                  };
              })
            | ({ id: number | null; uid: string } & {
                  " $fragmentRefs"?: {
                      ProductDetailsFragment_GroupedProduct_Fragment: ProductDetailsFragment_GroupedProduct_Fragment;
                  };
              })
            | ({ id: number | null; uid: string } & {
                  " $fragmentRefs"?: {
                      ProductDetailsFragment_SimpleProduct_Fragment: ProductDetailsFragment_SimpleProduct_Fragment;
                  };
              })
            | ({ id: number | null; uid: string } & {
                  " $fragmentRefs"?: {
                      ProductDetailsFragment_VirtualProduct_Fragment: ProductDetailsFragment_VirtualProduct_Fragment;
                  };
              })
            | null
        > | null;
    } | null;
};

type ProductDetailsFragment_BundleProduct_Fragment = {
    __typename: "BundleProduct";
    id: number | null;
    uid: string;
    meta_title: string | null;
    meta_keyword: string | null;
    meta_description: string | null;
    sku: string | null;
    name: string | null;
    categories: Array<{ uid: string; breadcrumbs: Array<{ category_uid: string } | null> | null } | null> | null;
    price: { regularPrice: { amount: { currency: CurrencyEnum | null; value: number | null } | null } | null } | null;
    small_image: { url: string | null } | null;
    description: { html: string } | null;
    short_description: { html: string } | null;
    media_gallery: Array<
        | { label: string | null; position: number | null; disabled: boolean | null; url: string | null }
        | { label: string | null; position: number | null; disabled: boolean | null; url: string | null }
        | null
    > | null;
} & { " $fragmentName"?: "ProductDetailsFragment_BundleProduct_Fragment" };

type ProductDetailsFragment_ConfigurableProduct_Fragment = {
    __typename: "ConfigurableProduct";
    id: number | null;
    uid: string;
    meta_title: string | null;
    meta_keyword: string | null;
    meta_description: string | null;
    sku: string | null;
    name: string | null;
    configurable_options: Array<{
        attribute_code: string | null;
        attribute_id: string | null;
        uid: string;
        label: string | null;
        values: Array<{
            uid: string | null;
            default_label: string | null;
            label: string | null;
            store_label: string | null;
            use_default_value: boolean | null;
            value_index: number | null;
            swatch_data:
                | { value: string | null }
                | { thumbnail: string | null; value: string | null }
                | { value: string | null }
                | null;
        } | null> | null;
    } | null> | null;
    categories: Array<{ uid: string; breadcrumbs: Array<{ category_uid: string } | null> | null } | null> | null;
    price: { regularPrice: { amount: { currency: CurrencyEnum | null; value: number | null } | null } | null } | null;
    small_image: { url: string | null } | null;
    description: { html: string } | null;
    short_description: { html: string } | null;
    media_gallery: Array<
        | { label: string | null; position: number | null; disabled: boolean | null; url: string | null }
        | { label: string | null; position: number | null; disabled: boolean | null; url: string | null }
        | null
    > | null;
} & { " $fragmentName"?: "ProductDetailsFragment_ConfigurableProduct_Fragment" };

type ProductDetailsFragment_DownloadableProduct_Fragment = {
    __typename: "DownloadableProduct";
    id: number | null;
    uid: string;
    meta_title: string | null;
    meta_keyword: string | null;
    meta_description: string | null;
    sku: string | null;
    name: string | null;
    categories: Array<{ uid: string; breadcrumbs: Array<{ category_uid: string } | null> | null } | null> | null;
    price: { regularPrice: { amount: { currency: CurrencyEnum | null; value: number | null } | null } | null } | null;
    small_image: { url: string | null } | null;
    description: { html: string } | null;
    short_description: { html: string } | null;
    media_gallery: Array<
        | { label: string | null; position: number | null; disabled: boolean | null; url: string | null }
        | { label: string | null; position: number | null; disabled: boolean | null; url: string | null }
        | null
    > | null;
} & { " $fragmentName"?: "ProductDetailsFragment_DownloadableProduct_Fragment" };

type ProductDetailsFragment_GroupedProduct_Fragment = {
    __typename: "GroupedProduct";
    id: number | null;
    uid: string;
    meta_title: string | null;
    meta_keyword: string | null;
    meta_description: string | null;
    sku: string | null;
    name: string | null;
    categories: Array<{ uid: string; breadcrumbs: Array<{ category_uid: string } | null> | null } | null> | null;
    price: { regularPrice: { amount: { currency: CurrencyEnum | null; value: number | null } | null } | null } | null;
    small_image: { url: string | null } | null;
    description: { html: string } | null;
    short_description: { html: string } | null;
    media_gallery: Array<
        | { label: string | null; position: number | null; disabled: boolean | null; url: string | null }
        | { label: string | null; position: number | null; disabled: boolean | null; url: string | null }
        | null
    > | null;
} & { " $fragmentName"?: "ProductDetailsFragment_GroupedProduct_Fragment" };

type ProductDetailsFragment_SimpleProduct_Fragment = {
    __typename: "SimpleProduct";
    id: number | null;
    uid: string;
    meta_title: string | null;
    meta_keyword: string | null;
    meta_description: string | null;
    sku: string | null;
    name: string | null;
    categories: Array<{ uid: string; breadcrumbs: Array<{ category_uid: string } | null> | null } | null> | null;
    price: { regularPrice: { amount: { currency: CurrencyEnum | null; value: number | null } | null } | null } | null;
    small_image: { url: string | null } | null;
    description: { html: string } | null;
    short_description: { html: string } | null;
    media_gallery: Array<
        | { label: string | null; position: number | null; disabled: boolean | null; url: string | null }
        | { label: string | null; position: number | null; disabled: boolean | null; url: string | null }
        | null
    > | null;
} & { " $fragmentName"?: "ProductDetailsFragment_SimpleProduct_Fragment" };

type ProductDetailsFragment_VirtualProduct_Fragment = {
    __typename: "VirtualProduct";
    id: number | null;
    uid: string;
    meta_title: string | null;
    meta_keyword: string | null;
    meta_description: string | null;
    sku: string | null;
    name: string | null;
    categories: Array<{ uid: string; breadcrumbs: Array<{ category_uid: string } | null> | null } | null> | null;
    price: { regularPrice: { amount: { currency: CurrencyEnum | null; value: number | null } | null } | null } | null;
    small_image: { url: string | null } | null;
    description: { html: string } | null;
    short_description: { html: string } | null;
    media_gallery: Array<
        | { label: string | null; position: number | null; disabled: boolean | null; url: string | null }
        | { label: string | null; position: number | null; disabled: boolean | null; url: string | null }
        | null
    > | null;
} & { " $fragmentName"?: "ProductDetailsFragment_VirtualProduct_Fragment" };